Fix the composer tripe for once and for all
I was running into problems where on my work machine, this was coming up
as ~/.composer; and on my laptop as ~/.config/composer.  I think they
changed at some point and put in 'smarts' to handle both.  But then I
barge in and fix it to one, and it didn't work on the other...

I know this is going to kill performance, but

1) It is one in the morning
2) I am the only person using my junk
3) I don't care anymore

AAAAARGH