Found another dumb hack: vm
I find myself moving things temporarily a lot (e.g. move an important
file somewhere to "break" some code to see how it handles that).
You'll eventually want to put it back.  You've typed the following:

    mv /path/to/very/important.file /path/to/very/important.file.bak

In a shell, how do you normally put things back?  If this were vim,
we'd probably take that old line, go to the first param, cut to the
next blank, and take that cut bit and paste it at the very end.  But
in a readline shell[1] this is an annoying process involving many
button presses, and the help of either screen/tmux or copy/paste in
your terminal.

Or, you could just define 'vm' which is 'mv' backwards.  Now all you
have to do is go to the beginning of your command, change 'mv' to
'vm' and BAM you're done.

[1] yes I know there's vim mode for shells including zsh... but let's
    just figure that I'm *really* used to the Emacs-style keys at a
    shell prompt and leave it at that