Remove $VISUAL editor
I already have my Rube Goldberg emacs setup wired up with a glorified
startup script[1], and I'd LIKE to do everything with that and not by
firing up my giant clunking Howl's Moving Castle of elisp every time I
need to visudo or crontab -e or whatever.

Problem is that with e.g. crontab -e, the calling program is waiting for
a return code, and em gives one immediately.  Which means the next best
solution is to just use vim for these tasks, and leave emacs for hacking
and org-mode.

[1] https://github.com/jws85/em